requirements.txt
connexion==1.1.14
Flask==0.12.2
flask-marshmallow==0.8.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
marshmallow-sqlalchemy==0.13.1
factory-boy==2.9.2
psycopg2==2.7.3.1
pytest-flask==0.10.0
